P3/File1/www/requirements.txt
flask
flask-mysqldb
mysql-connector-python
pycryptodomex
mysqlclient==1.3.14
flask-talisman